make the leds a shared sybsystem with implementation in v1_Gamma
